requirements.txt
anyio==3.7.1
APScheduler==3.10.1
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
gspread==5.10.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
oauthlib==3.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dotenv==1.0.0
python-telegram-bot==20.4
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
tzlocal==5.0.1
urllib3==1.26.16
